transitionning to header-only library
