+++ include/BaseGraph/directed_graph.hpp
@@ -1,5 +1,5 @@
-#ifndef BASE_GRAPH_LABELED_DIRECTED_GRAPH_H
-#define BASE_GRAPH_LABELED_DIRECTED_GRAPH_H
+#ifndef BASE_GRAPH_DIRECTED_GRAPH_HPP
+#define BASE_GRAPH_DIRECTED_GRAPH_HPP
 
 #include <iostream>
 #include <set>
