+++ python/CMakeLists.txt
@@ -16,7 +16,6 @@ find_package(pybind11 CONFIG REQUIRED)
 file(GLOB_RECURSE CORE_BINDINGS_SRC "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp")
 
 pybind11_add_module(_core MODULE ${CORE_BINDINGS_SRC})
-target_link_libraries(_core PRIVATE core)
 set_target_properties(_core PROPERTIES CXX_STANDARD 14)
 target_compile_definitions(_core PRIVATE VERSION_INFO=${PROJECT_VERSION})
 
@@ -24,6 +23,6 @@ message(STATUS "SKBUILD:" ${SKBUILD})
 if (SKBUILD)
     install(TARGETS _core DESTINATION ./basegraph)
 else()
-    set_target_properties(_core PROPERTIES 
+    set_target_properties(_core PROPERTIES
         LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/python-lib")
 endif()
