--- src/CMakeLists.txt
@@ -1,17 +1,0 @@
-set(BASEGRAPH_SRC undirected_multigraph.cpp directed_multigraph.cpp)
-
-add_library(core ${BASEGRAPH_SRC})
-set_target_properties(core PROPERTIES
-    LINKER_LANGUAGE CXX
-    POSITION_INDEPENDENT_CODE TRUE
-)
-if(MSVC)
-    set_target_properties(core PROPERTIES ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_ARCHIVE_OUTPUT_DIRECTORY} )
-    set_target_properties(core PROPERTIES ARCHIVE_OUTPUT_DIRECTORY_DEBUG ${CMAKE_ARCHIVE_OUTPUT_DIRECTORY} )
-    set_target_properties(core PROPERTIES ARCHIVE_OUTPUT_DIRECTORY_RELEASE ${CMAKE_ARCHIVE_OUTPUT_DIRECTORY} )
-endif(MSVC)
-
-target_include_directories(core PUBLIC
-    "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>"
-    "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>"
-)
