+++ tests/test_undirectedgraph.cpp
@@ -3,7 +3,7 @@
 #include <list>
 #include <stdexcept>
 
-#include "BaseGraph/edgelabeled_undirectedgraph.hpp"
+#include "BaseGraph/undirected_graph.hpp"
 #include "gtest/gtest.h"
 
 TEST(UndirectedGraph, getNeighboursOf_vertexOutOfRange_throwInvalidArgument) {
